refactor(T11404): reclaim core/src/tools — relocate brain-tools + task-tools dirs (E3 · SG-PACKAGE-ARCH)
Second reclaim wave (dir portion, 2 of 3). Moves the dir squatters up to honest
core siblings:

- core/src/tools/brain-tools/ → core/src/brain-tools/
- core/src/tools/task-tools/ → core/src/task-tools/

Each dir moved UP one level, so escaping relative imports were decremented
(../../ → ../); the tools/sdk/* files that import the task-tools sdk-tool were
repointed to ../../task-tools/; the tools/index.ts barrel re-exports from the
new locations. @cleocode/core public API is UNCHANGED (zero behavior change).

core/src/tools now holds only the atomic primitives (fs/shell/guard) + the sdk/
dir. The sdk/ reclaim (15 internal import fixups + 5 external importers incl a
contracts ref) is the final T11404 piece — T11404 stays open.

Verified: core build (tsc) · cleo build (public API intact) · cleo check arch
5/5 · deprecations + tools-vs-skills boundary clean · biome.
