+++ packages/core/src/tools/sdk/validator-ac-pull.ts
@@ -17,8 +17,8 @@
  */
 
 import { getTaskAccessor } from '../../store/data-accessor.js';
-import type { JsonSchema, RegisteredSdkTool } from '../task-tools/sdk-tool.js';
-import { defineSdkTool } from '../task-tools/sdk-tool.js';
+import type { JsonSchema, RegisteredSdkTool } from '../../task-tools/sdk-tool.js';
+import { defineSdkTool } from '../../task-tools/sdk-tool.js';
 
 /**
  * Input envelope for the {@link validatorAcPull} SDK tool.
